move all to package "bot"
