+++ src/main/scala/bot/dispatcher/Sheduler.scala
@@ -1,4 +1,4 @@
-package dispatcher
+package bot.dispatcher
 
 import java.util.concurrent.{Executors, TimeUnit}
