+++ src/main/scala/bot/telegram/api/Api.scala
@@ -1,36 +1,8 @@
-package telegram
-
-import org.json4s._
+package bot.telegram.api
 
 /**
-  * Implementation of telegram bot API.
-  * https://core.telegram.org/bots/api
+  * Created by lgor on 4/29/17.
   */
-
-trait TelegramBot {
-
-  implicit val formats = DefaultFormats
-
-  val token: String
-  val baseURL: String
-  val botURL: String
-
-  def getUpdates(): List[Update]
-
-  def sendMessage(chat_id: Int,
-                  text: String,
-                  reply_to_message_id: Option[Int] = None,
-                  parse_mode: Option[String] = None): Boolean
-}
-
-
-object TelegramBot {
-
-  def apply(token: String): TelegramBot = new APIImpl(token) with Logging
-}
-
-case class Response(ok: Boolean, description: Option[String], result: Option[List[Update]])
-
 case class User(id: Long, first_name: String, last_name: Option[String], username: Option[String])
 
 case class MessageEntity(`type`: String, offset: Int, length: Int, url: Option[String])
@@ -57,3 +29,6 @@ case class Chat(id: Int, `type`: String,
                 username: Option[String],
                 first_name: Option[String],
                 last_name: Option[String])
+
+
+case class Response(ok: Boolean, description: Option[String], result: Option[List[Update]])
